+++ VueForm/ConnectedCheckbox.js
@@ -24,10 +24,7 @@ export default {
     checked: String,
     indeterminate: String,
 
-    validate: {
-      type: Function,
-      default: noop,
-    },
+    validators: Array,
 
     handleFocus: {
       type: Function,
@@ -52,7 +49,7 @@ export default {
   data() {
     const $registerFormComponent = resolveRegisterFormComponent(this)
 
-    return $registerFormComponent(this.name, this.value, this.validate)
+    return $registerFormComponent(this.name, this.value, this.validators)
   },
 
   destroyed() {