+++ VueForm/ConnectedInputNumber.js
@@ -24,10 +24,7 @@ export default {
     controls: Boolean,
     controlsPosition: String,
 
-    validate: {
-      type: Function,
-      default: noop,
-    },
+    validators: Array,
 
     handleFocus: {
       type: Function,
@@ -53,7 +50,7 @@ export default {
   data() {
     const $registerFormComponent = resolveRegisterFormComponent(this)
 
-    return $registerFormComponent(this.name, this.value, this.validate)
+    return $registerFormComponent(this.name, this.value, this.validators)
   },
 
   destroyed() {