+++ VueForm/validations/isRequired.js
@@ -1,7 +1,0 @@
-import isNil from 'lodash/isNil'
-import startCase from 'lodash/startCase'
-
-export default function isRequired(customMessage) {
-  return (value, name) =>
-    isNil(value) ? customMessage || `${startCase(name)} is required` : undefined
-}