+++ VueForm/validators/index.js
@@ -1,9 +1,7 @@
 import isRequired from './isRequired'
 import length from './length'
-import validate from './validate'
 
 export default {
   isRequired,
   length,
-  validate,
 }